+++ lib/vanagon/cli/inspect.rb
@@ -1,5 +1,6 @@
 require 'docopt'
 require 'json'
+require 'vanagon/logger'
 
 class Vanagon
   class CLI
@@ -30,7 +31,7 @@ class Vanagon
       def parse(argv)
         Docopt.docopt(DOCUMENTATION, { argv: argv })
       rescue Docopt::Exit => e
-        puts e.message
+        VanagonLogger.error e.message
         exit 1
       end
 
@@ -41,7 +42,7 @@ class Vanagon
         platforms.each do |platform|
           driver = Vanagon::Driver.new(platform, project, options)
           components = driver.project.components.map(&:to_hash)
-          $stdout.puts JSON.pretty_generate(components)
+          VanagonLogger.warn JSON.pretty_generate(components)
         end
       end