+++ lib/vanagon/component/source/http.rb
@@ -1,4 +1,5 @@
 require 'vanagon/utilities'
+require 'vanagon/logger'
 require 'vanagon/component/source/local'
 require 'net/http'
 require 'uri'
@@ -93,7 +94,7 @@ class Vanagon
         #
         # @raise [RuntimeError] an exception is raised if the sum does not match the sum of the file
         def verify # rubocop:disable Metrics/AbcSize
-          warn "Verifying file: #{file} against sum: '#{sum}'"
+          VanagonLogger.info "Verifying file: #{file} against sum: '#{sum}'"
           actual = get_sum(File.join(workdir, file), sum_type)
           return true if sum == actual
 
@@ -107,7 +108,7 @@ class Vanagon
           uri = URI.parse(target_url.to_s)
           target_file ||= File.basename(uri.path)
 
-          warn "Downloading file '#{target_file}' from url '#{target_url}'"
+          VanagonLogger.info "Downloading file '#{target_file}' from url '#{target_url}'"
 
           Net::HTTP.start(uri.host, uri.port, use_ssl: uri.scheme == 'https') do |http|
             http.request(Net::HTTP::Get.new(uri, headers)) do |response|